Hacked in the fitness optimisation to UI

+++ src/main/java/com/thales/window/FitnessView.java
@@ -78,7 +78,7 @@ public class FitnessView extends Pane {
   }
 
 
-  public void addDataToQueue(int fitnessValue){
+  public void addDataToQueue(double fitnessValue){
     dataQ.add(fitnessValue);
   }
